Reduce peak memory during HuggingFace model import
- Load HF weights in bfloat16 instead of float32, halving the memory
  footprint of the temporary HF model.
- Delete the HF model immediately after extracting its state dict, and
  delete the raw HF state dict after mapping, so peak memory is closer
  to 1x model size instead of 2-3x.

+++ test_neural_net_model.py
@@ -989,7 +989,7 @@ class TestNeuralNetModel(unittest.TestCase):
 
         MockConfig.from_pretrained.assert_called_once_with("gpt2", revision="main")
         MockModel.from_pretrained.assert_called_once_with(
-            "gpt2", revision="main", torch_dtype=torch.float32, low_cpu_mem_usage=True
+            "gpt2", revision="main", dtype=torch.bfloat16, low_cpu_mem_usage=True
         )
